deprecate itemize/enumerateList actions
Use MessageUtils#itemizeList and MessageUtils#enumerateList. There is no reason to have these utility methods as actions. The only piece of logic used in the actions that hasn't been ported in the utility helper is the storage of the printed list in the session, but this was an experimental feature we aren't using anymore.